Create new ``used_before_assignment_typing`` test file
